Add skeleton files of a project
When starting development, sometimes basic things such as configuring
the language formatter (ie rustfmt) or deciding how to organize unit
tests are forgotten about until such time that they are needed. This
results in a lot of extra work and extra unnecessary commits cleaning up
the code base that could have been avoided (especially in the case of
the language formatter).

This change adds a basic skeleton of a project that sets up the
following:

- a .gitignore file that includes an entry for a top-level tmp/
  directory

- a rustfmt configuration file

- initial setup of unit test organization

This ensures that from the get go these basic things have already been
done reducing the needless churn of having to clean up the project later
in order to implement these things.