Add new struct: Path
The standard library Path type implements a mix of behaviour, some of
which involves calling out to the host operating system, and others
which do not. Since both types of behaviour are implemented on the same
type, there is a greater chance of performing a system operation
unintentionally which, depending on the operation and the authorization
given to the user running this code, may be disastrous.

This change adds a new Path struct that is intended to implement only
behaviour that does not depend on the host operating system. This
prevents any chance of accidentally performing any system operation; if
a system operation based on paths is needed, that is explicitly
requested by using a future type that specifically implements that
capability.