Add new trait: Attr
Any given path has specific attributes that are useful to know. However,
if access to path attributes is not available, then it becomes hard or
impossible to do things such as path validation or retrieving the parent
of a path.

This change adds a new trait, Attr, that will be defining all behaviour
that retrieves a specific attribute of a path. The specific attribute
implemented in this change is iterating over the individual components
that make up a path.