cargo: add new dependency: lazy_static
In order to implement platform specific behaviour, it is helpful to be
able to specify global constants using data structures such as a map or
a set. While the rust standard library includes such data structures, it
unfortunately does not include a way to use them as global constants.
This is due to the need of initializing and populating the data
structures which can only be done at run time.

This change adds a new dependency in Cargo.toml to the lazy_static
crate. This gives the ability to "use" data structures needing runtime
initialization/population as global constants.