Add new module: windows
The windows platform has its own specific properties and behaviour
regarding paths as compared to posix platforms. While some of thes
properties and behaviour can be found in the rust standard library, they
are not available for code compiles on a non-windows platform. This
means that windows paths cannot be created and/or manipulated easily on
a non-windows platform.

This change adds a new module containing windows platform properties and
behaviour, and is intended to be available regardless of the platform.
This change only establishes the module itself and defines some useful
constants that holds for all windows paths. Future commits will then be
focused on adding behaviour to this module.

Please note that this module will not contain any behaviour that calls
out to the system. The intent of this module is only to create and/or
manipulate paths, and not the system itself.