Add new windows struct: PathOps
The windows platform has behaviour for paths specific to the platform.
While this behaviour is available in the rust standard language, it is
only useable when compiled for the windows platform. As such, it is hard
to create and/or manipulate windows paths on a non-windows platform.

This change adds a new struct called PathOps. PathOps is intended to
contain behaviour regarding paths specific to the windows platform. By
putting this behaviour in a struct, it becomes easier to add this
behaviour where needed eg in a generic path handling struct that itself
needs to be able to deal with windows paths.

This change specifically only adds the PathOps struct and the
functions/methods needed to create an instance.

Please note that the PathOps struct will only contain behaviour for
dealing with pure paths. It will not contain any behaviour that requires
calling out to the system. This means, while the struct will be able to
create and manipulate paths, it will not contain behaviour to change
directories, validate paths exist, or any other behaviour requiring
system calls.