cargo: add new dependency: proptest
Creating unit tests is necessary in order to exercise a specific
functionality in order to verify that it behaves as expected under all
circumstances. For functionality that doesn't require any input or data,
this is relatively simple. Once input/data is needed, however, then it
gets complicated to properly test. All possible input should
ideally be tested, but it is expensive in terms of time spent to do so,
and attempting to do so manually has a high chance of forgetting to test
specific types of input ie edge/corner cases.

This change adds a new dev dependency in Cargo.toml on the proptest
library. proptest allows for the creation of property tests. These
property tests essentially automate the creation of input/data to feed
into a test. This means that all that needs to be specified is the
properties that a specific functionality has, and the proptest library
will then create all the tests needed based on the expected input the
functionality is expected to retrieve. This greatly reduces the
amount of time needed to create a reasonable set of tests, and it also
decreases the chance of forgetting to test edge/corner cases.