Remove module: windows
The windows module was implemented by porting code from python's pathlib
module. However, the resulting code is hard to understand and test. This
was demonstrated by the difficulty in creating easy to write unit tests.
As such, there is concern that this module is not maintainable.

This change removes the windows module. Future commits will recreate
this module with a from-scratch implementation.