Add new module: windows::match_prefix
The windows platform has several different types of paths. Each path is
more or less differentiated by their different prefixes. In order to
parse a windows path, it is needed to parse the path's prefix if it
exists.

This change adds a new module, `match_prefix`, that parses the prefix of
a given windows path.

Please note: this specific commit depends on commit 8830649c1041 ("Add
new module: windows::path_type").