Add new struct: CurrentDir
Even though it is valid to specify the current directory via a `.` for a
path component, there is no type that validates this. Without a way to
specify the current directory, otherwise valid paths may generate an
error when iterating over path components.

This change adds the CurrentDir struct that is only equal to the byte
string b".";