Add new struct: ParentDir
Even though it is valid to specify the parent directory via a `..` for a
path component, there is n o type that validates this. Without such a
type, otherwise valid paths may generate an error when iterating over
path components.

This change adds the ParentDir struct that is only equal to the byte
string b"..".