Add new structs: ValidLastChar and InvalidLastChar
Path component names on a windows platform cannot end with a space or a
period character. However, the current path component name validators do
are unaware of this restriction, and there is no way to check for
invalid last character. As a result, invalid paths may be considered
valid.

This change adds 2 structs, `ValidLastChar` and `InvalidLastChar`, that
are meant to be used to check if the last character of a path component
is valid.

Please note that this change does not include usage of these 2 new
structs which will be added in future commits.