Improve NonUNCPart validation
The NonUNCPart type is the component that comes after `//?/` for paths
with a Verbatim prefix. However, this type allows this component to be
named `.` or `..` or a device name. This is invalid since this component
is still considered part of the prefix and not yet part of the path
proper.

This change ensures that the Verbatim prefix cannot be named `.`, `..`,
or a device name.