fix: NonDevicePart allows names ending in `.` or a space
The NonDevicePart type is any component of the path ie any component
after the root component. However, this type allows the component's name
to end with a `.` or a space character even though this is invalid as
per the Windows file system name specification.

This change updates the NonDevicePart type so that it returns false when
comparing against a name that ends with a `.` or a space character.