fix: device names w/ extension not recognized as devices
A name usually matches as a device using the full name. However, the
windows platform also specifies that if a file name has an extension,
if the name before the extension matches a device name then that file is
considered a device ie any extension is ignored when determining if a
file name matches a device name. However, the implementation of the
Device type only checks against the full file name; if the file name
includes an extension, the extension is included w/ the check.

This change ensures that for any name that has an extension, the
extension is ignored before checking to see if the name matches a device
name.