Use Result as PathIterator's item
If a path component cannot be parsed due to an invalid character, the
iterator will return a component item error. However, this is merely an
variant of the Component enum type and not a real error. Because of
this, code that is running the PathIterator will not be able to use the
normal error handling facilities (ie ?/try!) making it harder to process
errors returned from the iterator.

This change changes the item used by the PathIterator type to be a
Result instead of a bare Component enum variant. This ensures that code
running the iterator will easily be able to detect and handle errors
returned by the iterator.