fix: PathIterator error msg states invalid range
When iterating over a PathIterator object, if an error is returned, the
error message states the start and end indexes in the path string that
corresponds to the component that generated an error. However, the
indexes stated in the message are incorrect and do not correspond to the
correct path component. The cause of this issue is due to assigning the
index of the next component before generating the error; error
generating depends on the index of the current component, and with the
index being set to the next component, any error message will always
state the wrong indices.

This change ensures that the index stored during iteration is not
updated to the start of the next path component until the current
component has been processed. With this change, if an error is
generated, then the error message will state the correct start and end
indices of the component that generated the error.