Add new error kind generated by PathIterator
When an error is generated during iteration over a PathIterator, the
error is always due to a parsing error. However, it is not specified
what type of parse error occurred. As such, when processing the error,
it is difficult to decide what to do with the error as only a generic
message is provided.

This change defines a second type of error when iterating over a
PathIterator. The 2 types of errors that can be generated are:

1. Path component contains 1 or more invalid characters

2. Path component name is a restricted name