fix: path starting with . or .. are considered invalid
Iterating over a path that starts with either . (current dir) or ..
(parent dir) stops iteration and yields an error. This is an issue as
such paths are actually valid. This is caused since both of these paths
end with a . and path components are not allowed to end with a . (except
for . and ..).

This change updates path component validation so that . or .. are not
considered invalid.