fix: NonDevicePart test uses . and ..
The unit test called `ne_invalid_lastchar` allows using `.` and `..` as
filenames when checking for an invalid last character in the filename.
Using these values causes the test to fail since they are not invalid
filenames.

This change ensures that `.` and `..` are never chosen as values to test
in `ne_invalid_lastchar`.