Refactor PartialEq impl for windows::path_type::Device
The `PartialEq` impl for the `Device` struct uses the
UnicodeSegmentation crate in order to not check against codepoints that
are in the middle of a unicode character. This is done in order to find
where a file extension starts in a string. However, the implementation
does not always find the file extension when it exists.

This change removes usage of the UnicodeSegmentation crate in the
`PartialEq` impl for the `Device` struct, and in re-implementing ensures
that the position of the file extension is correctly identified.