Reduce max size of proptest generated vecs
Many vectors that are randomly generated by the `proptest` library have
a maximum length of 256. Although tests successfully run and pass, it
takes a noticeable length of time to do so for tests that generate such
vectors.

This change changes the maximum length of these vectors to 10; 256 was
just an arbitrary number that's meaningless as nothing in the code
tested cares about length of the vectors much. With this change, tests
run much faster.