Add script that implements tarpaulin workaround
When cargo-tarpaulin is run against a project that has the proptest
library as a dependency, it triggers an internal compiler error. Until
the compiler error is fixed, it is not possible to use tarpaulin by
itself to calculate code coverage.

This change adds a script that allows tarpaulin to be run against a
project with a proptest dependency. It does this by disabling the
linking against dead code for the proptest library only.