Add justfile with commands to calculate code coverage
In order to calculate code coverage, some setup is required in order for
code coverage tools to run without error. However, doing so is tedious
as it requires doing many things such as setting environment variables
and ensuring certain directories exist.

This change adds a justfile to be used by the just cli tool. This tool
is an alternative to the make cli tool. Included in the justfile are
commands to calculate code coverage using either cargo-tarpaulin or
kcov, where each command performs all needed setup in order for either
code coverage tool to be able to run.