Use mk_reverse_equal!() macro from common module
There are 2 copies of the mk_reverse_equal!() macro: the original
defined in the windows::path_type module, and another defined in
crate::common.

This change removes the macro defined in windows::path_type and uses the
one defined in crate::common instead. This removes repeated code and
also makes it so that the new unix module will not depend on the windows
module.

This change depends on commit 6e2012cdf621 ("Copy mk_reverse_equal!()
into common module").