Copy CurrentDir and ParentDir impl into common module
The unix module defines types used for matching path components. This
includes '.' for the current directory and '..' for the parent
directory. However, these types have already been defined in the windows
module. If importing from the windows module, then the unix module will
have a non-sensical dependency on the windows module. The other
alternative is to copy these structs to the unix module at the cost of
violating DRY principles.

This change copies the CurrentDir and ParentDir structs into the
common::path_type module. Future commits will then remove the
corresponding types from the windows module and use the types from the
common module instead.