Use CurrentDir and ParentDir types from common module
There are 2 copies of both the CurrentDir and ParentDir types: the
original types defined in the windows::path_type module, and the others
defined in the common::path_type module.

This change removes the types defined in the windows::path_type module
and uses the types defined in the common::path_type module instead. This
removes repeated code and also makes it so that the new unix module will
not depend on the windows module.

Depends on commit bb2bb5066de0 ("Copy CurrentDir and ParentDir impl into
common module").