Move unit tests from windows::path_type to common::path_type
The unit tests for the CurrentDir and ParentDir types are still in the
windows::path_type module. This is confusing since these types are now
defined in the common::path_type module.

This change moves these unit tests to the common::path_type module where
the CurrentDir and ParentDir types are defined.