Add new unix type: Part
In order to properly parse unix style paths, a path component needs to
be able to be verified if it is valid or not.

This change adds a new type called `Part` that matches against any set
of unicode characters as long as neither the separator character `/` nor
the null character are included in the set.