Implement PartialEq<u8> for unix::path_type::Separator
When using the Separator type to check if a character is a separator,
the check is done against a slice. However, the separator character, if
valid, is always a single u8 value making it cumbersome to create a
slice for just a single character.

This change improves the ergonomics of using the Separator type to match
against a character by adding an implementation of PartialEq<u8>. This
means that instead of creating a slice, a byte value make get extracted
directly and used for the comparison. This makes it easier to perform
the comparison.