Remove unix::path_type::Part type
The unix::path_type::Part type is used to determine if a byte slice is a
valid unix path file or directory name. In order to do this, it iterates
over every byte of the slice. However, this is already being done in the
PathIterator type when trying to determine if a byte slice contains
separator characters. Thus, for a given slice, the slice is being
evaluated multiple times.

This change removes the unix::path_type::Part type and moves the invalid
character check into the PathIterator type. By doing this, each byte of
a byte slice is only evaluated once.