Copy error types into new common::error module
The error types between the unix and windows modules are more or less
the same with the only differences being in the enum used for the error.

This change copies the ParseError and related types to a new module
`common::error`.

Please note, that as part of this change, that new enums are added to
the unix and windows modules that essentially duplicates their
respective ParseErrorKind enum types. This is only temporary until the
modules are changes in future commits to use the unified error in the
common::error module.