Copy string helper functions into new module common::string
The string handler functions used when parsing a path are exactly the
same in the unix and windows and is duplicated code.

This change adds a copy of these functions into a new module
`common::string`. Future commits will then use functions in
common::string instead of the copies found in the unix and windows
modules.