Use string functions from common::string in unix module
The string functions used in the unix module are duplicated code of
identical functions in the windows module.

This change removes the string functions from the unix module and
replaces them with functions imported from the common::string module.
Once the windows module is similarly updated, there should be no
duplicated code for these string functions.