Use string functions from common::string in windows module
The string functions used in the windows module are duplicated code of
identical functions in the unix module.

This change removes the string functions from the windows module and
replaces them with functions imported from the common::string module.
With this change, there should be no duplicated code for these string
functions.