fix: iterating over empty windows path yields nothing
If the windows path used for the iterator is an empty string, then
nothing is yielded from the iterator. It is expected, though, that the
empty string refers to the current directory.

This change fixes the issue by explicitly handling the empty string case
when iterating through the path; an empty string will always yield a
single item which is the current directory.