Add Path struct
The stdlib Path struct provides a lot of functionality, however it is
gated by the platform it is being run on. This means that one cannot
manipulate windows paths on a unix platform or vice versa.

The Path struct implemented in this crate will make a distinction
between a memory only path and a system path allowing some path
operations regardless of platform.

This change implements only the base Path struct and a few relevant
AsRef traits. Future commits will flesh out the functionality.