Implement AsRef traits for types in unix and windows modules
The Component and PathIterator types in the unix and windows modules are
not able to be directly converted into a Path type. As a result, doing
this will require more boilerplate than is necessary.

This change implements the requisite AsRef traits for the Component and
PathIterator types of the unix and windows modules so that these types
are able to be converted into a Path type. As a result, it is much
easier to use these types to create new Path instances.