Rename PathIterator -> Iter
The iterator type over a path's component parts is called
`PathIterator`. This name, however, may be confusing to some as it is
sounds ambiguous what the it iterates over eg path components, or paths?

This change renames the PathIterator type to simply the Iter type.
Combined with the expected import usage (eg use pathlib::unix::Iter),
this removes any ambiguity over what the iterator iterates over,
especially as it is solely used with the Path concrete type.