Add new trait: MemoryPath
Path and PathBuf types in the standard library implement all possible
operations. However, due to this, manipulating platform paths are
restricted to the respective only. This means, when on a unix platform
windows paths cannot be used, and vice versa.

In order to make it so that at least non-system operations are permitted
regardless of platform, operations need to be separated into system and
non-system categories.

This change creates a new trait called MemoryPath which is what will
categorize non-system operations. Included in this change is a macro
that implements MemoryPath for a specific type given an iterator.

Please note that this change also creates a marker trait called
ComponentIterator which simply marks an iterator type that iterates over
path components. This trait is needed for the MemoryPath trait to work.