Add new macro: path_struct
Path types need to be implemented specific to the platform the
code in running on. However, the majority of the implementation is the
same regardless of platform. In order to create the platform specific
implementations, a lot of duplication will be done.

This change removes the duplication by adding a new macro called
`path_struct`. This macro defines and implements the parts of a Path
struct that are platform agnostic. Any implementation of the Path type
will then only need to implement the platform specific bits.