Use platform specific Path and PathBuf structs in path module
The implementation of the Path and PathBuf structs depend on
platform-specific details; it is not possible to have just a single
implementation of these 2 structs without some form of conditional
compilation based on platform the code is running on. Thus, the
implementation of the Path struct in the path module is insufficient
when considering both unix and windows platforms due to differences in
how file names are encoded.

This change removes the implementation of the Path struct from the path
module and instead imports the Path and PathBuf structs from the
appropriate platform module. This ensures that the structs best fit to
the platform the code is running on is used.