Update Iter contructor based on platform
The platform specific Iter type accepts only a byte slice as an argument
to its constructor. However, on the windows platform, it is unable to
meaningfully create such a byte slice when creating an Iter type;
filenames on the windows platform are encoded in a UTF-16 like encoding
which has no UTF-8 analogue. As such, the only way to create a list of
bytes representing a filename on the windows platform is to allocate
memory in an owned type such as a Vec.

This change uses conditional compilation to define the Iter types based
on the platform the code is running on. On unix, the Iter type's
constructor will continue to accept a byte slice; on windows, the Iter
type's constructor will instead only accept a Vec<u8>.

Included in this change are ensuring methods defined for the Iter type
are able to operate in the same way whether a byte slice or a Vec<u8>
was used to create the type, and ensuring current unit tests that create
an Iter type are aware of platform differences.