Implement MemoryPath trait for unix path types
When wanting to manipulate paths for a specific platform, it is
difficult to navigate the various modules and import the needed traits
and types and fairly unergonomic.

This change defines and implements the UnixMemoryPath and
UnixMemoryPathBuf traits for the Path and PathBuf unix path types. With
this trait, users simply have to import the unix specific MemoryPath
trait in order to use unix platform path handling.