Add macro to implement generic memory path traits
The idea to use marker traits to simplify implementation of a common
interface between platform specific implementations of path types didn't
work out due to, on the unix platform, the size of types being unknown
in the unix::Iter struct since the struct stores a reference to a byte
slice. In order to fix this, the MemoryPath trait will be removed and
its behaviour folded into platform specific memory path traits.

This change adds a macro to define and implement the generic interface
of platform specific memory path traits.