Replace unix MemoryPath implementation with UnixMemoryPath
The idea to use marker traits to simplify implementation of a common
interface between platform specific implementations of path types didn't
work out due to, on the unix platform, the size of types being unknown
in the unix::Iter struct since the struct stores a reference to a byte
slice. In order to fix this, the MemoryPath trait will be removed and
its behaviour folded into platform specific memory path traits.

This change removes from the unix module the implementation of the now
obsolete MemoryPath trait, and in its place changes the UnixMemoryPath
trait to define the interface that the MemoryPath trait had. With this
change, the unix module no longer depends on the MemoryPath trait and
its associated macro `impl_memorypath`.