Replace windows MemoryPath implementation with WindowsMemoryPath
The idea to use marker traits to simplify implementation of a common
interface between platform specific implementations of path types didn't
work out due to, on the unix platform, the size of types being unknown
in the unix::Iter struct since the struct stores a reference to a byte
slice. In order to fix this, the MemoryPath trait will be removed and
its behaviour folded into platform specific memory path traits.

Although windows::Iter stores an owned type and doesn't have the type
size issue at compile time, to keep maintainability high, the windows
platform path types should be implemented in the same way that the unix
platform path types are.

This change removes from the windows module the implementation of the
now obsolete MemoryPath trait, and in its place changes the
WindowsMemoryPath trait to define the interface that the MemoryPath
trait had. With this change, the windows module no longer depends on the
MemoryPath trait and its associated macro `impl_memorypath`.