Remove MemoryPath and associated traits
The idea to use marker traits to simplify implementation of a common
interface between platform specific implementations of path types didn't
work out due to, on the unix platform, the size of types being unknown
in the unix::Iter struct since the struct stores a reference to a
byte slice.

In order to fix this, this change removes the MemoryPath and
ComponentIterator traits.