Export PathBuf type from crate root
The most used types in this crate are the Path and PathBuf types.
However, the PathBuf type is not re-exported in the crate root. This
makes it a little bit unergonomic to use the type.

This change exports the PathBuf type from the crate root, making it
easier to import the most commonly used types.