Implement memory path for borrowed type
When using a memory path trait to specify which interface to use,
passing in a reference to a Path type fails to compile. This is due to
the memory path not being implemented on a reference to the Path type.

This change implements memory path traits for not only the concrete Path
and PathBuf types but also references to these types.