Re-export std::path::Prefix from windows module
When importing types needed to use the windows memory path trait, the
Prefix enum is not included. Instead, this enum must be imported from
the standard library. This is not very ergonomic and is harder to
discover.

This change re-exports std::path::Prefix from the windows module. This
makes it easier to simply import the needed types from seemingly the
same crate.