Add a prelude module
The prelude module is meant to be a simple way to import all necessary
types and traits for the majority of the use cases. However, this module
is not implemented for this crate. This means it is slightly more
cumbersome to get started as a user will need to know which modules to
import from to just get started with paths for the running platform.

This change adds a prelude module. The module exports path types
defaulting to the platform the code is running on. This makes it easier
to quickly get started using paths for the running platform.