Change implementation of Iter struct
The design of the Iter struct for both unix path and windows path
iteration was such that it only really worked when running on a unix
platform; on windows there is constantly issues regarding lifetimes.
This is due to OsStr not being able to expose its underlying bytes
on the windows platform resulting in always having to create owned
objects when manipulating OS strings.

This commit changes the implementation of the Iter struct for both unix
and windows path iteration so that it accepts a reference to a Path.

Please note that in order to re-implement the Iter struct, a few changes
were needed to be made to the Path and PathBuf types. Now, all types
behave accordingly based on the platform the code is running on (eg Path
and PathBuf types use &[u8] on a unix platform, while these types must
use Vec<u16> on a windows platform).