fix: coverage report including files for different platform
The tarpaulin coverage tool often includes files that should not be
counted in coverage reports eg modules that will only be compiled on
the windows platform are being counted when tests are running on a unix
platform.

This change excludes all modules that are only compiled into the build
artifact when on a windows platform. Please note that this is valid for
now as the tarpaulin tool is not able to run on windows.