Remove code coverage script workaround
When running the tarpaulin code coverage tool, an internal compiler
error is raised. A workaround exists consisting of setting specific
python script as the rustc compiler to use when running tarpaulin with
the script turning off dead-code generation for the proptest crate.
However, use of a python script is not very portable and depends on
python being installed.

Since v0.9.1, the proptest crate includes a workaround itself which is
enabled behind a feature flag. Since this is a much less hacky
workaround, this change removes the python script and instead enables
the `default-code-coverage` feature of proptest v0.9.1.