Implement support for windows platform
While it was always the intention to be able to work on both unix and
windows platforms, initial development could only be done on the unix
platform. As a result, many issues exist that prevent the library from
working on the windows platform. For example: syntax errors, difference
in handling of file names, unit tests written only from the unix
platform perspective.

This change updates the library so that both the library code and the
unit test suite work as expected on the windows platform.