Add traits representing memory paths
A memory path is simply a path that exists in memory only ie does not
make any system calls.

The current implementation of path types creates traits representing the
behaviour of paths on a per-platform basis. This means that there are
separate and distinct traits for the unix and windows platforms. In
order to be useful, these separate traits essentially defines the exact
same methods, however this makes maintenance harder as any change to the
interface of one platform trait must also then be applied to the
corresponding trait for the other platform.

This change defines a two new traits `MemoryPath` and `MemoryPathBuf`.
These traits is intended to be platform agnostic, and as such, the
interface of paths is generalized. This results in an interface that is
better able to be maintained.