Add new type: UnixPath
The current implementation of path types creates traits representing the
behaviour of paths on a per-platform basis. This means that there are
separate and distinct traits for the unix and windows platforms. In
order to be useful, these separate traits essentially defines the exact
same methods, however this makes maintenance harder as any change to the
interface of one platform trait must also then be applied to the
corresponding trait for the other platform. It is also more confusing to
use as having both unix and windows paths in the same scope requires
calling path methods via a specific trait ie
UnixMemoryPath::path_method() or WindowsMemoryPath::path_method().

This change defines two new types `UnixPath` and `UnixPathBuf` that
implements the new traits `MemoryPath` and `MemoryPathBuf`. This makes
it easier to use especially when unix paths and windows paths are in the
same scope (or will be once concrete windows paths have been
implemented).

This change depends on commit 644761af8e56 ("Add traits representing
memory paths").