Remove obsolete path traits
With the implementation of concrete path types, the old path traits are
no longer needed; the new MemoryPath and MemoryPathBuf traits along with
the UnixPath, UnixPathBuf, WindowsPath, and WindowsPathBuf concrete
types completely replace the functionality in the UnixMemoryPath,
UnixMemoryPathBuf, WindowsMemoryPath, and WindowsMemoryPathBuf traits.

This change removes the old path traits so that there is no duplication
of functionality in the library.