Replace macro path_struct with Path struct
The path_struct macro is used to implement a Path struct in the same way
on unix and windows platforms. However, the core implementation of the
struct is exactly the same for both platforms. This means that the macro
itself is unneeded.

This change removes the path_struct macro and replaces it with an
implementation of the Path type. Platform specific implementatinos
remain unchanged and now use the single new Path type instead of
defining the type for each platform.