Replace use of unsafe blocks w/ safe functions
Some methods use unsafe blocks in order to convert bytes into strings or
os strings. This can be hard to audit and maintain since these blocks
are scattered throughout the codebase.

This change replaces use of unsafe blocks with calls to safe functions
that wrap the unsafe functionality. This makes it easier to keep track
of, audit, and maintain unsafe usage as the unsafety is in a single
module ie crate::common::string.

Please note that there is other usage of unsafe blocks in crate::path,
but this is also easy enough to keep track of as it is confined to only
a single method along with the implementation of the Sync and Send
marker traits.