fix: integration tests fail on windows platform
When running the integration tests on the windows platform, several
tests fail. This is due to the windows tests being only a copy of the
unix tests; since the behaviour of the library on the windows platform
is slightly different than the unix behaviour, the tests that pass on
unix fail on windows.

This change updates the tests so that they account for the differences
in how the library works when being run on the windows platform.