Add azure pipeline build config
While this library is developed on the unix platform, the library is
meant to be cross-platform ie it works on both unix and windows
platforms. However, unit tests are only being run on the unix platform.
As a result, it is very easy for bugs either in the library code or in
the unit and/or integration tests to be added.

This change adds a configuration for building and running the unit tests
via the azure pipeline service. Part of this configuration is running
the tests on the windows platform. This should make it easier to
determine if bugs are added when running on the windows platform.