Add travisci build config
While this library is developed on nightly rust, it is intended to work
on the stable and beta versions as well. This change adds a
configuration for building and running the unit tests on stable, beta,
and nightly rust via the travisci service.

Please note that the travisci config is for linux only. Please see the
azure pipelines config file for windows and mac os builds.