Add new utility function: os_str_to_bytes()
On the windows platform, there is no way to get a reference to the byte
array underlying the OsStr type; the only facility provided is to
allocate a vec of u16 integers representing UTF-16 ish codepoints.
However, without a way to retrieve a reference to the underlying bytes,
the custom windows iterator is not able to be used due to lifetime
issues.

This change adds a function that provides a safe wrapper around an
unsafe cast from &OsStr to *const [u8]. While nothing in this commit
uses it, future commits will be use this function to implement getting a
reference to the underlying byte array of an OsStr when on the windows
platform.

This function is considered safe to use as it provides read-only access
to the underlying byte array. And so, as warned in the rust stdlib docs,
the returned byte array is not exactly UTF-8 and so may contain
"codepoints" that are invalid. However, for the purposes of parsing a
path string, this is perfectly acceptable as the only characters being
checked for are all within the 7-bit ascii range.