Implement Path::as_bytes() when on windows platform
On the windows platform, there is no way to get a reference to the byte
array underlying the OsStr type; the only facility provided is to
allocate a vec of u16 integers representing UTF-16 ish codepoints via
the OsStrExt trait.  However, without a way to retrieve a reference to
the underlying bytes, the custom windows iterator is not able to be used
due to lifetime issues.

This change uses the os_str_to_bytes() function added in commit 2810ad7fabef
("Add new utility function: os_str_to_bytes()") to implement the
Path::as_bytes() method when on the windows platform. This will allow
using the same windows iterator struct used to parse windows file
paths on the unix platform when on the windows platform.

Using the same iterator struct on the windows platform as on the unix
platform is desired since the current windows platform implementation
simply uses the iterator provided by the stdlib which has slightly
different behaviour. This leads to having to split the unit tests by
platform introducing a lot of duplicated test code. By using the same
iterator, tests can eventually be made platform agnostic removing the
need for the platform split in the unit tests.