Use unix impl of windows path iterator for windows platform
The implementation of an iterator over the path components of a windows
path behaves different on the windows platform than on the unix
platform. This is because due to not being able to get a reference to
the underlying byte array of the OsStr type, the rust stdlib
implementation of the iterator is used. However, this leads to a bit of
a maintenance burden since there is now 2 different behaviours for
parsing a windows path depending on platform.

This change copies the unix platform implementation of a windows path
component iterator and uses it for the windows platform implementation.
This makes it much easier to maintain going forward as there is now only
a single expected behaviour and it is consistent regardless of platform.

Please note that this change includes updating the unit test suite to
account for the behaviour of the iterator when on the windows platform
being the same as on the unix platform.